requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pypdf>=5.0.0
streamlit>=1.36.0
